CM-792: Refactor filters
