+++ src/gatsby/src/pages/find-a-park.js
@@ -3,9 +3,6 @@ import { graphql, Link as GatsbyLink } from "gatsby"
 import axios from "axios"
 import { orderBy } from "lodash"
 import {
-  Checkbox,
-  FormGroup,
-  FormControlLabel,
   Chip,
   TextField,
   InputAdornment,
@@ -17,8 +14,6 @@ import {
 import ClearIcon from '@mui/icons-material/Clear'
 import SearchIcon from "@mui/icons-material/Search"
 import CancelIcon from "@mui/icons-material/Cancel"
-import ExpandLess from "@mui/icons-material/ExpandLess"
-import ExpandMore from "@mui/icons-material/ExpandMore"
 import OpenInNewIcon from '@mui/icons-material/OpenInNew'
 
 import Footer from "../components/footer"
@@ -26,7 +21,8 @@ import Header from "../components/header"
 import Seo from "../components/seo"
 import ScrollToTop from "../components/scrollToTop"
 import NoSearchResults from "../components/search/noSearchResults"
-import SearchFilter from "../components/search/searchFilter"
+import MobileFilters from "../components/search/mobileFilters"
+import DesktopFilters from "../components/search/desktopFilters"
 import ParkLinksModal from "../components/search/parkLinksModal"
 
 import "../styles/search.scss"
@@ -95,50 +91,12 @@ export const query = graphql`
   }
 `
 
-const Filter = ({ filterItems, selectedFilterItems, handleFilterCheck }) => {
-  return (
-    <FormGroup className="filter-options-container">
-      {filterItems.map(item =>
-        <FormControlLabel
-          key={item.label}
-          control={
-            <Checkbox
-              checked={
-                selectedFilterItems.filter(
-                  selectedFilterItem =>
-                    selectedFilterItem.value === item.value
-                ).length === 1 ? true : false
-              }
-              onChange={event => {
-                handleFilterCheck(item, event)
-              }}
-              name={item.label}
-            />
-          }
-          label={`${item.label} (${item.count})`}
-          className={
-            selectedFilterItems.filter(
-              selectedFilterItem =>
-                selectedFilterItem.value === item.value
-            ).length === 1 ? "text-light-blue no-wrap" : "no-wrap"
-          }
-          disabled={item.count === 0}
-        />
-      )}
-    </FormGroup>
-  )
-}
-
 export default function FindAPark({ location, data }) {
   const menuContent = data?.allStrapiMenu?.nodes || []
 
   const [regionsCount, setRegionsCount] = useState([])
   const [activitiesCount, setActivitiesCount] = useState([])
   const [facilitiesCount, setFacilitiesCount] = useState([])
-  const [showMoreActivities, setMoreActivites] = useState(true)
-  const [showMoreFacilities, setMoreFacilities] = useState(true)
-  const [truncatedActivityFilterLength, setTruncatedActivityFilterLength] = useState(5)
-  const [truncatedFacilityFilterLength, setTruncatedFacilityFilterLength] = useState(5)
 
   const sortedActivityItems = orderBy(
     data.allStrapiActivityType.nodes,
@@ -458,19 +416,6 @@ export default function FindAPark({ location, data }) {
     currentPage
   ])
 
-  useEffect(() => {
-    if (showMoreActivities) {
-      setTruncatedActivityFilterLength(5)
-    } else {
-      setTruncatedActivityFilterLength(activityItems.length)
-    }
-    if (showMoreFacilities) {
-      setTruncatedFacilityFilterLength(5)
-    } else {
-      setTruncatedFacilityFilterLength(facilityItems.length)
-    }
-  }, [showMoreActivities, activityItems.length, showMoreFacilities, facilityItems.length])
-
   return (
     <>
       <Header content={menuContent} />
@@ -722,118 +667,29 @@ export default function FindAPark({ location, data }) {
                     </div>
                     <div className="col-12 pr-3 mb32">
                       <h3 className="subtitle mb-2">Filter</h3>
-                      <div className="">
-                        <fieldset className="mb-2">
-                          <legend className="filter-heading p10t">Popular</legend>
-                          <Filter
-                            filterItems={campingFacilityItems.filter(
-                              c => c.value === 36
-                            )}
-                            selectedFilterItems={selectedCampingFacilities}
-                            handleFilterCheck={handleCampingFacilityCheck}
-                          />
-                          <Filter
-                            filterItems={activityItems.filter(
-                              a => a.value === 1 || a.value === 8 || a.value === 9
-                            )}
-                            selectedFilterItems={selectedActivities}
-                            handleFilterCheck={handleActivityCheck}
-                          />
-                          <Filter
-                            filterItems={facilityItems.filter(f => f.value === 6)}
-                            selectedFilterItems={selectedFacilities}
-                            handleFilterCheck={handleFacilityCheck}
-                          />
-                          <Filter
-                            filterItems={activityItems.filter(
-                              a => a.value === 3
-                            )}
-                            selectedFilterItems={selectedActivities}
-                            handleFilterCheck={handleActivityCheck}
-                          />
-                          <Filter
-                            filterItems={campingFacilityItems.filter(
-                              c => c.value === 1
-                            )}
-                            selectedFilterItems={selectedCampingFacilities}
-                            handleFilterCheck={handleCampingFacilityCheck}
-                          />
-                        </fieldset>
-                        <fieldset className="mb-2">
-                          <legend className="filter-heading">Regions</legend>
-                          <Filter
-                            filterItems={regionItems}
-                            selectedFilterItems={selectedRegions}
-                            handleFilterCheck={handleRegionCheck}
-                          />
-                        </fieldset>
-                        <fieldset className="mb-2">
-                          <legend className="filter-heading">Camping</legend>
-                          <Filter
-                            filterItems={campingFacilityItems}
-                            selectedFilterItems={selectedCampingFacilities}
-                            handleFilterCheck={handleCampingFacilityCheck}
-                          />
-                        </fieldset>
-                        <fieldset className="mb-2">
-                          <legend className="filter-heading">Activities</legend>
-                          <Filter
-                            filterItems={activityItems.slice(0, truncatedActivityFilterLength)}
-                            selectedFilterItems={selectedActivities}
-                            handleFilterCheck={handleActivityCheck}
-                          />
-                          <Link
-                            className="ml-auto pointer"
-                            onClick={() => {
-                              setMoreActivites(!showMoreActivities)
-                            }}
-                            tabIndex="0"
-                            role="link"
-                            underline="hover"
-                          >
-                            {showMoreActivities ? (
-                              <div style={{ color: `#2464A4` }}>
-                                Show all {activityItems.length}
-                                <ExpandMore fontSize="small" />
-                              </div>
-                            ) : (
-                              <div style={{ color: `#2464A4` }}>
-                                Show less
-                                <ExpandLess fontSize="small" />
-                              </div>
-                            )}
-                          </Link>
-                        </fieldset>
-                        <fieldset>
-                          <legend className="filter-heading">Facilities</legend>
-                          <Filter
-                            filterItems={facilityItems.slice(0, truncatedFacilityFilterLength)}
-                            selectedFilterItems={selectedFacilities}
-                            handleFilterCheck={handleFacilityCheck}
-                          />
-                          <Link
-                            className="ml-auto pointer"
-                            onClick={() => {
-                              setMoreFacilities(!showMoreFacilities)
-                            }}
-                            tabIndex="0"
-                            role="link"
-                            underline="hover"
-                          >
-                            {showMoreFacilities ? (
-                              <div style={{ color: `#2464A4` }}>
-                                Show all {facilityItems.length}
-                                <ExpandMore fontSize="small" />
-                              </div>
-                            ) : (
-                              <div style={{ color: `#2464A4` }}>
-                                Show less
-                                <ExpandLess fontSize="small" />
-                              </div>
-                            )}
-                          </Link>
-                        </fieldset>
-                      </div>
+                      <DesktopFilters
+                        data={{
+                          regionItems,
+                          campingFacilityItems,
+                          activityItems,
+                          facilityItems,
+                          selectedRegions,
+                          setSelectedRegions,
+                          selectedCampingFacilities,
+                          setSelectedCampingFacilities,
+                          selectedActivities,
+                          setSelectedActivities,
+                          selectedFacilities,
+                          setSelectedFacilities,
+                          searchText,
+                          setCurrentPage,
+                          setFilters,
+                          handleRegionCheck,
+                          handleCampingFacilityCheck,
+                          handleActivityCheck,
+                          handleFacilityCheck
+                        }}
+                      />
                     </div>
                     <div className="col-12 park-links">
                       <h3 className="subtitle mb-2">More ways to find a park</h3>
@@ -892,7 +748,7 @@ export default function FindAPark({ location, data }) {
           </div>
         </div>
       </div>
-      <SearchFilter
+      <MobileFilters
         data={{
           totalResults,
           regionItems,
@@ -911,6 +767,11 @@ export default function FindAPark({ location, data }) {
           setSelectedFacilities,
           searchText,
           setCurrentPage,
+          setFilters,
+          handleRegionCheck,
+          handleCampingFacilityCheck,
+          handleActivityCheck,
+          handleFacilityCheck
         }}
       />
       <ParkLinksModal data={{ openModal, setOpenModal }} />
